Step 2/2 - renaming mobileWebView to the proper mobileWebViewUtils - updating references to the file with the new name
