+++ src/components/common/version.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import packageJson from '../../../package.json';
-import { isMobileWebView} from '../../services/utilities/mobileWebView';
+import { isMobileWebView} from '../../services/utilities/mobileWebViewUtils';
 
 
 export default class Version extends Component {
   render() {
     const version = <span style={{fontSize: "10px",float:"right"}}>v{packageJson.version} </span>;
-    
+
     if (isMobileWebView) {
       return (
         <div>
